refactor: move socket based test server implementation behind abstraction
in order to be able to introduce new, http based server for testing as well.
