+++ thrifty-test-server/src/main/java/com/microsoft/thrifty/testing/TestServerInterface.java
@@ -1,0 +1,11 @@
+package com.microsoft.thrifty.testing;
+
+import com.microsoft.thrifty.test.gen.ThriftTest;
+
+public interface TestServerInterface {
+    void run(ServerProtocol protocol, ServerTransport transport);
+
+    int port();
+
+    void close();
+}
